frontend/requirements.txt
gradio==4.1.1
grpcio==1.59.0